backend/requirements.txt
Django>=4.2
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pandas>=2.0.0
reportlab>=4.0.0
django-rest-framework-simplejwt>=5.2.0
